+++ .env-example
@@ -1,3 +1,3 @@
 CHANNEL=chrome
 HEADLESS=False
-PASSWORD=
+PASSWORD=<DEFAULT_PASSWORD_FOR_METAMASK>
